// convex/_generated/api.d.ts
/* eslint-disable */
/**
 * Generated `api` utility.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * To regenerate, run `npx convex dev`.
 * @module
 */

import type {
  ApiFromModules,
  FilterApi,
  FunctionReference,
} from "convex/server";
import type * as accommodations from "../accommodations.js";
import type * as activities from "../activities.js";
import type * as auth from "../auth.js";
import type * as http from "../http.js";
import type * as itinerary from "../itinerary.js";
import type * as notes from "../notes.js";
import type * as packing from "../packing.js";
import type * as reminders from "../reminders.js";
import type * as router from "../router.js";
import type * as trips from "../trips.js";

/**
 * A utility for referencing Convex functions in your app's API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
declare const fullApi: ApiFromModules<{
  accommodations: typeof accommodations;
  activities: typeof activities;
  auth: typeof auth;
  http: typeof http;
  itinerary: typeof itinerary;
  notes: typeof notes;
  packing: typeof packing;
  reminders: typeof reminders;
  router: typeof router;
  trips: typeof trips;
}>;
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;